+++ rocks/__init__.py
@@ -3,10 +3,19 @@
 # Welcome to rocks
 __version__ = "1.6.4"
 
-# Expose API to user
-from .core import Rock
-from .core import rocks_ as rocks
-from .resolve import identify
+import inspect
 
-# Alias id to identify
-id = identify
+if "rocks.cli" not in inspect.stack()[-1].code_context[0]:
+
+    # Expose API to user
+    from .core import Rock
+    from .core import rocks_ as rocks
+    from .resolve import identify
+
+    # Alias id to identify
+    id = identify
+
+
+# print(globals())
+# print("click" in globals())
+# print("CLICKED" in globals())